p.225

+++ chapter03/microservices/review-service/src/main/java/se/magnus/microservices/core/review/services/ReviewServiceImpl.java
@@ -1,19 +1,25 @@
 package se.magnus.microservices.core.review.services;
 
 import java.util.List;
+import java.util.function.Supplier;
 
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
 
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import se.magnus.api.core.review.Review;
 import se.magnus.api.core.review.ReviewService;
 import se.magnus.microservices.core.review.persistence.ReviewEntity;
 import se.magnus.microservices.core.review.persistence.ReviewRepository;
 import se.magnus.util.exceptions.InvalidInputException;
 import se.magnus.util.http.ServiceUtil;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
@@ -23,29 +29,40 @@ public class ReviewServiceImpl implements ReviewService {
     private final ReviewRepository repository;
     private final ReviewMapper mapper;
 	private final ServiceUtil serviceUtil;
+	private final Scheduler scheduler;
     
 	@Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+		this.scheduler = scheduler;
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
     }
 	
 	@Override
-	public List<Review> getReviews(int productId) {
+	public Flux<Review> getReviews(int productId) {
 		if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
-		
+
+		LOG.info("Will get reviews for product with id={}", productId);
+
+		return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, FINE);		
+	}
+	
+	protected List<Review> getByProductId(int productId) {
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
         list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
-		
-        LOG.debug("/reviews response size: {}", list.size());
         
+        LOG.debug("/reviews response size: {}", list.size());
+
         return list;
 	}
 
 	@Override
 	public Review createReview(Review body) {
+		if (body.getProductId() < 1) 
+			throw new InvalidInputException("Invalid productId: " + body.getProductId());
+		
         try {
             ReviewEntity entity = mapper.apiToEntity(body);
             ReviewEntity newEntity = repository.save(entity);
@@ -54,15 +71,20 @@ public class ReviewServiceImpl implements ReviewService {
             return mapper.entityToApi(newEntity);
 
         } catch (DataIntegrityViolationException dive) {
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Review Id:" + body.getReviewId());
+            throw new InvalidInputException("Duplicate Key, Product Id: " + body.getProductId() + ", Review Id:" + body.getReviewId());
         }
 	}
 
 	@Override
 	public void deleteReviews(int productId) {
+		
+		if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
 		LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
-
         repository.deleteAll(repository.findByProductId(productId));
+	}
+	
+	private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+		return Flux.defer(publisherSupplier).subscribeOn(scheduler);
 	}
 }
